+++ docker/openam/am-pre-start-script.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+# This is a placeholder that you can replace with your own script to copy in assets such as images
+# or a web.xml file.
+# Environment variables:
+# CATALINA_HOME: The expanded openam war file is in $CATALINA_HOME/webapps/openam
+# OPENAM_HOME
+echo "AM pre-start script"
+
+env
+
+